XD-3130 Create BusCleanerUtils
 - Move `buildRestTemplate` method from `BusUtils` to a new class `BusCleanerUtils`
 - Since `Spark-streaming` uses `BusUtils`, this change would make Spark streaming not
to depend on `httpClient` that is used by the bus cleaner util method to build rest template
